+++ WeaveManila-Project/src/pages/Owner_Module/Dashboard/InventoryModule/InventoryPages/SegregatePage.vue
@@ -680,9 +680,10 @@ bordered
         }
       },
       checkUserStatus() {
-          axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
-          .then(response => {
-          const latestStatus = response.data.information.status;
+        axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
+        .then(response => {
+          console.log(response.data);
+
           const information = response.data.information;
             this.information = {
               id: information.id,
@@ -705,25 +706,29 @@ bordered
               isAdmin: information.isAdmin,
             };
             SessionStorage.set('information', JSON.stringify(this.information));
-          // Update the local status and take appropriate action if it has changed
-          if (this.status !== latestStatus) {
-            this.status = latestStatus;
 
-            if (this.status === 0) {
-              this.$q.notify({
-                type: 'negative',
-                message: 'Your account is currently inactive. Please contact the administrator.',
-              });
-              this.$router.push('/');
-              sessionStorage.clear();
+            const latestAdmin = response.data.information.isAdmin;
+            const latestStatus = response.data.information.status;
+
+            if (this.status !== latestStatus || (latestAdmin && latestStatus === 0)) {
+              this.status = latestStatus;
+
+              if (this.status === 0 && !latestAdmin) {
+                this.$q.notify({
+                  type: 'negative',
+                  message: 'Your account is currently inactive. Please contact the administrator.',
+                });
+                this.$router.push('/');
+                sessionStorage.clear();
+              }
             }
-          }
         }).catch(error => {
               console.error('Error fetching data:', error);
         });
       },
       loadUserData() {
         const userData = SessionStorage.getItem('information');
+
         if (userData) {
           try {
             const userInformation = JSON.parse(userData);
@@ -736,10 +741,10 @@ bordered
             this.position = userInformation.position;
             this.status = userInformation.status;
             this.isAdmin = userInformation.isAdmin;
-
             this.fullname = this.firstname + " " + this.lastname;
             if (this.position.toLowerCase() === 'owner') {
-              this.$router.push('/dashboard/segregate-section');
+
+              this.$router.push('/dashboard/main-dashboard');
             } else {
 
               this.$q.notify({
@@ -750,8 +755,7 @@ bordered
               sessionStorage.clear();
             }
 
-            if (this.status == 0) {
-
+            if (this.status === 0 && !this.isAdmin) {
               this.$q.notify({
                 type: 'negative',
                 message: 'Your account is currently inactive. Please contact the administrator.',
@@ -759,6 +763,7 @@ bordered
               this.$router.push('/');
               sessionStorage.clear();
             }
+
           } catch (error) {
             console.log('Error parsing user data:', error);
             // Provide user feedback or navigate to an error page
