+++ WeaveManila-Project/src/pages/Owner_Module/Dashboard/InventoryModule/MpoForm.vue
@@ -1370,9 +1370,10 @@ export default {
       }
     },
     checkUserStatus() {
-        axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
-        .then(response => {
-        const latestStatus = response.data.information.status;
+      axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
+      .then(response => {
+        console.log(response.data);
+
         const information = response.data.information;
           this.information = {
             id: information.id,
@@ -1395,19 +1396,22 @@ export default {
             isAdmin: information.isAdmin,
           };
           SessionStorage.set('information', JSON.stringify(this.information));
-        // Update the local status and take appropriate action if it has changed
-        if (this.status !== latestStatus) {
-          this.status = latestStatus;
 
-          if (this.status === 0) {
-            this.$q.notify({
-              type: 'negative',
-              message: 'Your account is currently inactive. Please contact the administrator.',
-            });
-            this.$router.push('/');
-            sessionStorage.clear();
+          const latestAdmin = response.data.information.isAdmin;
+          const latestStatus = response.data.information.status;
+
+          if (this.status !== latestStatus || (latestAdmin && latestStatus === 0)) {
+            this.status = latestStatus;
+
+            if (this.status === 0 && !latestAdmin) {
+              this.$q.notify({
+                type: 'negative',
+                message: 'Your account is currently inactive. Please contact the administrator.',
+              });
+              this.$router.push('/');
+              sessionStorage.clear();
+            }
           }
-        }
       }).catch(error => {
             console.error('Error fetching data:', error);
       });
@@ -1427,11 +1431,10 @@ export default {
           this.position = userInformation.position;
           this.status = userInformation.status;
           this.isAdmin = userInformation.isAdmin;
-
           this.fullname = this.firstname + " " + this.lastname;
           if (this.position.toLowerCase() === 'owner') {
 
-            this.$router.push('/dashboard/mpoform-section');
+            this.$router.push('/dashboard/main-dashboard');
           } else {
 
             this.$q.notify({
@@ -1442,8 +1445,7 @@ export default {
             sessionStorage.clear();
           }
 
-          if (this.status == 0) {
-
+          if (this.status === 0 && !this.isAdmin) {
             this.$q.notify({
               type: 'negative',
               message: 'Your account is currently inactive. Please contact the administrator.',
@@ -1451,7 +1453,9 @@ export default {
             this.$router.push('/');
             sessionStorage.clear();
           }
+
         } catch (error) {
+          console.log('Error parsing user data:', error);
           // Provide user feedback or navigate to an error page
           this.$q.notify({
             type: 'negative',
