+++ WeaveManila-Project/src/pages/Owner_Module/Dashboard/UserManagement/UserManagement.vue
@@ -1001,12 +1001,31 @@ export default {
       this.showEditArea = false;
     },
     // Old Data
+    toggleProduction(){
+      this.productionVisible = !this.productionVisible;
+    },
+    toggleInventoryMenu() {
+      this.inventoryMenuVisible = !this.inventoryMenuVisible;
+    },
+    toggleDrawer() {
+      if (!this.toggleDrawers) {
+        this.drawer = true;
+        this.drawerWidth = 300;
+        this.drawerIcon = 'arrow_forward_ios';
+        this.toggleDrawers = true;
+      } else {
+        this.drawer = true;
+        this.drawerWidth = 80;
+        this.toggleDrawers = false;
+        this.drawerIcon = 'arrow_back_ios_new';
+      }
+    },
     checkUserStatus() {
-        axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
-        .then(response => {
-        const latestStatus = response.data.information.status;
+      axios.get(`http://localhost/Capstone-Project/backend/api/verification.php?email=${this.email}`)
+      .then(response => {
+        console.log(response.data);
+
         const information = response.data.information;
-        const Position = response.data.information.position;
           this.information = {
             id: information.id,
             email: information.email,
@@ -1028,50 +1047,25 @@ export default {
             isAdmin: information.isAdmin,
           };
           SessionStorage.set('information', JSON.stringify(this.information));
-        if (Position.toLowerCase() === 'owner') {
-          this.$router.push('/dashboard/usermanagement-section');
-        }else{
-          this.$q.notify({
-          type: 'negative',
-            message: 'You do not have permission to access the system.',
-          });
-          this.$router.push('/');
-          sessionStorage.clear();
-        }
-        if (this.status !== latestStatus) {
-          this.status = latestStatus;
 
-          if (this.status === 0) {
-            this.$q.notify({
-              type: 'negative',
-              message: 'Your account is currently inactive. Please contact the administrator.',
-            });
-            this.$router.push('/');
-            sessionStorage.clear();
+          const latestAdmin = response.data.information.isAdmin;
+          const latestStatus = response.data.information.status;
+
+          if (this.status !== latestStatus || (latestAdmin && latestStatus === 0)) {
+            this.status = latestStatus;
+
+            if (this.status === 0 && !latestAdmin) {
+              this.$q.notify({
+                type: 'negative',
+                message: 'Your account is currently inactive. Please contact the administrator.',
+              });
+              this.$router.push('/');
+              sessionStorage.clear();
+            }
           }
-        }
       }).catch(error => {
             console.error('Error fetching data:', error);
       });
-    },
-    toggleProduction(){
-      this.productionVisible = !this.productionVisible;
-    },
-    toggleInventoryMenu() {
-      this.inventoryMenuVisible = !this.inventoryMenuVisible;
-    },
-    toggleDrawer() {
-      if (!this.toggleDrawers) {
-        this.drawer = true;
-        this.drawerWidth = 300;
-        this.drawerIcon = 'arrow_forward_ios';
-        this.toggleDrawers = true;
-      } else {
-        this.drawer = true;
-        this.drawerWidth = 80;
-        this.toggleDrawers = false;
-        this.drawerIcon = 'arrow_back_ios_new';
-      }
     },
     loadUserData() {
       const userData = SessionStorage.getItem('information');
@@ -1088,27 +1082,29 @@ export default {
           this.position = userInformation.position;
           this.status = userInformation.status;
           this.isAdmin = userInformation.isAdmin;
-          this.fullnames = this.firstname + " " + this.lastname;
+          this.fullname = this.firstname + " " + this.lastname;
           if (this.position.toLowerCase() === 'owner') {
-            this.$router.push('/dashboard/usermanagement-section');
-          }else{
+
+            this.$router.push('/dashboard/main-dashboard');
+          } else {
+
             this.$q.notify({
-            type: 'negative',
+              type: 'negative',
               message: 'You do not have permission to access the system.',
             });
             this.$router.push('/');
             sessionStorage.clear();
           }
-          if(this.status == 0)
-          {
+
+          if (this.status === 0 && !this.isAdmin) {
             this.$q.notify({
-            type: 'negative',
+              type: 'negative',
               message: 'Your account is currently inactive. Please contact the administrator.',
             });
             this.$router.push('/');
             sessionStorage.clear();
           }
-          // If another value add here
+
         } catch (error) {
           console.log('Error parsing user data:', error);
           // Provide user feedback or navigate to an error page
